HC IndividualBlogHome 코드 리팩토링
