+++ blogfront/src/components/IndividualBlogHome/TempCategory.js
@@ -1,0 +1,3 @@
+const list = ['유럽 여행', '아시아 여행', '오세아니아 여행', '아메리카 여행']
+
+export default list;
