+++ blogfront/src/components/IndividualBlogHome/TempPostName.js
@@ -1,0 +1,18 @@
+const list = [
+
+    {category: "유럽 여행", date: "2023.05.02~2023.05.13", title: "영국, 아일랜드, 프랑스, 스위스, 오스트리아, 체코"},
+    {category: "유럽 여행", date: "2023.01.12~2023.02.03", title: "아랍에미레이트, 영국, 스페인, 프랑스, 스위스, 이탈리아"},
+    {category: "유럽 여행", date: "2024.03.02~2024.03.15", title: "영국, 스위스, 독일"},
+
+    {category: "아시아 여행", date: "2010.10.10~2010.10.13", title: "일본"},
+    {category: "아시아 여행", date: "2015.07.12~2015.07.17", title: "필리핀"},
+    {category: "아시아 여행", date: "2016.12.12~2016.12.16", title: "싱가포르"},
+    {category: "아시아 여행", date: "2016.08.25~2019.08.31", title: "대만"},
+
+    {category: "오세아니아 여행", date: "2007.10.12~2007.10.22", title: "뉴질랜드"},
+
+    {category: "아메리카 여행", date: "2011.02.03~2011.03.02", title: "미국, 캐나다"}
+
+]
+
+export default list;
